GitHub Action을 사용한 자동 CICD 워크플로우 추가 : feat : 폴더 구조 및 dto 스타일 변경https://github.com/TEAM-ROMROM/RomRom-BE/issues/4
