+++ build.gradle
@@ -59,9 +59,6 @@ dependencies {
     // Redis
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 
-    // Validation
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-
     // swagger
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.3'
 }
