+++ src/main/java/com/romrom/romback/domain/controller/MemberController.java
@@ -1,0 +1,5 @@
+package com.romrom.romback.domain.controller;
+
+public class MemberController {
+
+}
