--- src/main/java/com/romrom/romback/domain/member/controller/MemberController.java
@@ -1,5 +1,0 @@
-package com.romrom.romback.domain.member.controller;
-
-public class MemberController {
-
-}
