+++ src/main/java/com/romrom/romback/domain/service/MemberService.java
@@ -1,13 +1,12 @@
-package com.romrom.romback.domain.member.service;
+package com.romrom.romback.domain.service;
 
-import com.romrom.romback.domain.member.domain.AccountStatus;
-import com.romrom.romback.domain.member.domain.Member;
-import com.romrom.romback.domain.member.domain.Role;
-import com.romrom.romback.domain.authentication.dto.SignUpRequest;
-import com.romrom.romback.domain.member.repository.MemberRepository;
+import com.romrom.romback.domain.object.constant.AccountStatus;
+import com.romrom.romback.domain.object.postgres.Member;
+import com.romrom.romback.domain.object.constant.Role;
+import com.romrom.romback.domain.object.dto.AuthRequest;
+import com.romrom.romback.domain.repository.postgres.MemberRepository;
 import com.romrom.romback.global.exception.CustomException;
 import com.romrom.romback.global.exception.ErrorCode;
-import com.romrom.romback.global.util.BaseResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -26,10 +25,9 @@ public class MemberService {
    * 회원가입
    *
    * @param request username, password, nickname
-   * @return 없음
    */
   @Transactional
-  public BaseResponse<Void> signUp(SignUpRequest request) {
+  public Void signUp(AuthRequest request) {
 
     // 사용자 아이디 검증 (중복 아이디 사용 불가)
     if (memberRepository.existsByUsername(request.getUsername())) {
@@ -47,7 +45,7 @@ public class MemberService {
     );
     log.debug("회원가입 성공: username={}", request.getUsername());
 
-    return BaseResponse.success(null);
+    return null;
   }
