+++ src/main/java/com/romrom/romback/global/config/SecurityUrls.java
@@ -8,27 +8,26 @@ import java.util.List;
  */
 public class SecurityUrls {
 
-    /**
-     * 인증을 생략할 URL 패턴 목록
-     */
-    public static final List<String> AUTH_WHITELIST = Arrays.asList(
-            // API
-            "/api/auth/signup", // 회원가입
+  /**
+   * 인증을 생략할 URL 패턴 목록
+   */
+  public static final List<String> AUTH_WHITELIST = Arrays.asList(
+      // API
+      "/api/auth/signup", // 회원가입
 
+      // Swagger
+      "/docs/**", // Swagger UI
+      "/v3/api-docs/**" // Swagger API 문서
 
-            // Swagger
-            "/docs/**", // Swagger UI
-            "/v3/api-docs/**" // Swagger API 문서
+  );
 
-    );
+  /**
+   * 관리자 권한이 필요한 URL 패턴 목록
+   */
+  public static final List<String> ADMIN_PATHS = Arrays.asList(
+      // API
+      "/admin/concert-hall/**"
 
-    /**
-     * 관리자 권한이 필요한 URL 패턴 목록
-     */
-    public static final List<String> ADMIN_PATHS = Arrays.asList(
-            // API
-            "/admin/concert-hall/**"
-
-    );
+  );
 
 }
