+++ src/main/java/com/romrom/romback/global/config/SwaggerConfig.java
@@ -14,53 +14,53 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @OpenAPIDefinition(
-        info = @Info(
-                title = "📱 롬롬 : ROM-ROM 📱",
-                description = """
-                        ### 🌐 롬롬 웹사이트 🌐 : romrom.co.kr
-                        [**웹사이트 바로가기**](http://suh-project.synology.me:8085)
+    info = @Info(
+        title = "📱 롬롬 : ROM-ROM 📱",
+        description = """
+            ### 🌐 롬롬 웹사이트 🌐 : romrom.co.kr
+            [**웹사이트 바로가기**](http://suh-project.synology.me:8085)
 
-                        ### 💻 **GitHub 저장소**
-                        - **[백엔드 소스코드](https://github.com/TEAM-ROMROM/RomRom-BE)**
-                          백엔드 개발에 관심이 있다면 저장소를 방문해보세요.
-                        """,
-                version = "1.0v"
-        ),
-        servers = {
-                @Server(url = "http://suh-project.synology.me:8085", description = "메인 서버"),
-                @Server(url = "http://suh-project.synology.me:8086", description = "테스트 서버"),
-                @Server(url = "http://localhost:8080", description = "로컬 서버")
-        }
+            ### 💻 **GitHub 저장소**
+            - **[백엔드 소스코드](https://github.com/TEAM-ROMROM/RomRom-BE)**
+              백엔드 개발에 관심이 있다면 저장소를 방문해보세요.
+            """,
+        version = "1.0v"
+    ),
+    servers = {
+        @Server(url = "http://suh-project.synology.me:8085", description = "메인 서버"),
+        @Server(url = "http://suh-project.synology.me:8086", description = "테스트 서버"),
+        @Server(url = "http://localhost:8080", description = "로컬 서버")
+    }
 )
 @Configuration
 public class SwaggerConfig {
 
-    @Bean
-    public OpenAPI openAPI() {
-        SecurityScheme apiKey = new SecurityScheme()
-                .type(Type.HTTP)
-                .in(In.HEADER)
-                .name("Authorization")
-                .scheme("bearer")
-                .bearerFormat("JWT");
+  @Bean
+  public OpenAPI openAPI() {
+    SecurityScheme apiKey = new SecurityScheme()
+        .type(Type.HTTP)
+        .in(In.HEADER)
+        .name("Authorization")
+        .scheme("bearer")
+        .bearerFormat("JWT");
 
-        SecurityRequirement securityRequirement = new SecurityRequirement()
-                .addList("Bearer Token");
+    SecurityRequirement securityRequirement = new SecurityRequirement()
+        .addList("Bearer Token");
 
-        return new OpenAPI()
-                .components(new Components().addSecuritySchemes("Bearer Token", apiKey))
-                .addSecurityItem(securityRequirement)
-                .servers(List.of(
-                                new io.swagger.v3.oas.models.servers.Server()
-                                        .url("http://localhost:8080")
-                                        .description("로컬 서버"),
-                                new io.swagger.v3.oas.models.servers.Server()
-                                        .url("http://suh-project.synology.me:8086")
-                                        .description("테스트 서버"),
-                                new io.swagger.v3.oas.models.servers.Server()
-                                        .url("http://suh-project.synology.me:8085")
-                                        .description("메인 서버")
-                        )
-                );
-    }
+    return new OpenAPI()
+        .components(new Components().addSecuritySchemes("Bearer Token", apiKey))
+        .addSecurityItem(securityRequirement)
+        .servers(List.of(
+                new io.swagger.v3.oas.models.servers.Server()
+                    .url("http://localhost:8080")
+                    .description("로컬 서버"),
+                new io.swagger.v3.oas.models.servers.Server()
+                    .url("http://suh-project.synology.me:8086")
+                    .description("테스트 서버"),
+                new io.swagger.v3.oas.models.servers.Server()
+                    .url("http://suh-project.synology.me:8085")
+                    .description("메인 서버")
+            )
+        );
+  }
 }
