+++ src/main/java/com/romrom/romback/global/exception/ErrorResponse.java
@@ -1,12 +1,14 @@
 package com.romrom.romback.global.exception;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
+@Builder
 public class ErrorResponse {
 
-    private ErrorCode errorCode;
-    private String errorMessage;
+  private ErrorCode errorCode;
+  private String errorMessage;
 }
