+++ src/main/java/com/romrom/romback/global/filter/TokenAuthenticationFilter.java
@@ -1,7 +1,7 @@
 package com.romrom.romback.global.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.romrom.romback.domain.member.service.CustomUserDetailsService;
+import com.romrom.romback.domain.service.CustomUserDetailsService;
 import com.romrom.romback.global.config.SecurityUrls;
 import com.romrom.romback.global.exception.ErrorCode;
 import com.romrom.romback.global.exception.ErrorResponse;
@@ -28,137 +28,137 @@ import org.springframework.web.filter.OncePerRequestFilter;
 @Slf4j
 public class TokenAuthenticationFilter extends OncePerRequestFilter {
 
-    private final JwtUtil jwtUtil;
-    private final CustomUserDetailsService customUserDetailsService;
-    private static final AntPathMatcher pathMatcher = new AntPathMatcher();
+  private final JwtUtil jwtUtil;
+  private final CustomUserDetailsService customUserDetailsService;
+  private static final AntPathMatcher pathMatcher = new AntPathMatcher();
 
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws ServletException, IOException {
+  @Override
+  protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
+      throws ServletException, IOException {
 
-        String uri = request.getRequestURI();
-        log.debug("요청된 URI: {}", uri);
+    String uri = request.getRequestURI();
+    log.debug("요청된 URI: {}", uri);
 
-        // 화이트리스트 체크 : 화이트리스트 경로면 필터링 건너뜀
-        if (isWhitelistedPath(uri)) {
-            filterChain.doFilter(request, response);
-            return;
+    // 화이트리스트 체크 : 화이트리스트 경로면 필터링 건너뜀
+    if (isWhitelistedPath(uri)) {
+      filterChain.doFilter(request, response);
+      return;
+    }
+
+    // 요청 타입 구분 : API 요청/관리자 페이지 요청
+    boolean isApiRequest = uri.startsWith("/api/");
+    boolean isAdminPageRequest = uri.startsWith("/admin/");
+
+    try {
+      String token = null;
+      String bearerToken = request.getHeader("Authorization");
+      // 토큰 추출: 요청 타입에 따라 헤더 또는 파라미터에서 토큰 추출
+      if (isApiRequest) {
+        log.debug("일반 API 요청입니다.");
+        // API 요청 : Authorization 헤더에서 "Bearer " 토큰 추출
+        if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
+          token = bearerToken.substring(7).trim(); // "Bearer " 제거
+        }
+      } else if (isAdminPageRequest) {
+        log.debug("관리자 페이지 요청입니다.");
+        // 관리자 페이지 요청: Authorization 헤더 우선 확인
+        if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
+          token = bearerToken.substring(7).trim();
+        } else {
+          // Authorization 헤더에 토큰이 없는 경우 파라미터 확인
+          String paramToken = request.getParameter("accessToken");
+          if (paramToken != null && !paramToken.isEmpty()) {
+            token = paramToken;
+          }
+        }
+      }
+
+      // 토큰 검증: 토큰이 유효하면 인증 설정
+      if (token != null && jwtUtil.validateToken(token)) {
+        Authentication authentication = jwtUtil.getAuthentication(token);
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+
+        // 관리자 페이지 접근 권한 체크: 관리자 권한 없으면 로그인 페이지로 리다이렉트
+        if (isAdminPageRequest && !hasAdminRole(authentication)) {
+          log.error("관리자 권한이 없습니다. 로그인페이지로 리다이렉트합니다.");
+          response.sendRedirect("/login");
+          return;
         }
 
-        // 요청 타입 구분 : API 요청/관리자 페이지 요청
-        boolean isApiRequest = uri.startsWith("/api/");
-        boolean isAdminPageRequest = uri.startsWith("/admin/");
-
-        try {
-            String token = null;
-            String bearerToken = request.getHeader("Authorization");
-            // 토큰 추출: 요청 타입에 따라 헤더 또는 파라미터에서 토큰 추출
-            if (isApiRequest) {
-                log.debug("일반 API 요청입니다.");
-                // API 요청 : Authorization 헤더에서 "Bearer " 토큰 추출
-                if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
-                    token = bearerToken.substring(7).trim(); // "Bearer " 제거
-                }
-            } else if (isAdminPageRequest) {
-                log.debug("관리자 페이지 요청입니다.");
-                // 관리자 페이지 요청: Authorization 헤더 우선 확인
-                if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
-                    token = bearerToken.substring(7).trim();
-                } else {
-                    // Authorization 헤더에 토큰이 없는 경우 파라미터 확인
-                    String paramToken = request.getParameter("accessToken");
-                    if (paramToken != null && !paramToken.isEmpty()) {
-                        token = paramToken;
-                    }
-                }
-            }
-
-            // 토큰 검증: 토큰이 유효하면 인증 설정
-            if (token != null && jwtUtil.validateToken(token)) {
-                Authentication authentication = jwtUtil.getAuthentication(token);
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-
-                // 관리자 페이지 접근 권한 체크: 관리자 권한 없으면 로그인 페이지로 리다이렉트
-                if (isAdminPageRequest && !hasAdminRole(authentication)) {
-                    log.error("관리자 권한이 없습니다. 로그인페이지로 리다이렉트합니다.");
-                    response.sendRedirect("/login");
-                    return;
-                }
-
-                // 인증 성공
-                filterChain.doFilter(request, response);
-                return;
-            }
-
-            // 토큰이 없거나 유효하지 않은 경우
-            if (isApiRequest) {
-                // 토큰 없음
-                if (token == null) {
-                    log.error("토큰이 존재하지 않습니다.");
-                    sendErrorResponse(response, ErrorCode.MISSING_AUTH_TOKEN);
-                } else { // 유효하지 않은 토큰
-                    log.error("토큰이 유효하지 않습니다.");
-                    sendErrorResponse(response, ErrorCode.INVALID_ACCESS_TOKEN);
-                }
-                return; // 필터 체인 진행하지 않음
-            } else if (isAdminPageRequest) {
-                // 관리자 페이지: 로그인 페이지로 리다이렉트
-                log.error("관리자 페이지 요청 시, 토큰이 없거나 유효하지 않습니다.");
-                response.sendRedirect("/login");
-                return;
-            }
-        } catch (ExpiredJwtException e) {
-            log.error("토큰 만료: {}", e.getMessage());
-            // 토큰 만료 예외 처리
-            if (isApiRequest) {
-                sendErrorResponse(response, ErrorCode.EXPIRED_ACCESS_TOKEN);
-            } else {
-                response.sendRedirect("/login");
-            }
-            return;
-        }
-
-        // 필터 체인 계속 진행
+        // 인증 성공
         filterChain.doFilter(request, response);
+        return;
+      }
+
+      // 토큰이 없거나 유효하지 않은 경우
+      if (isApiRequest) {
+        // 토큰 없음
+        if (token == null) {
+          log.error("토큰이 존재하지 않습니다.");
+          sendErrorResponse(response, ErrorCode.MISSING_AUTH_TOKEN);
+        } else { // 유효하지 않은 토큰
+          log.error("토큰이 유효하지 않습니다.");
+          sendErrorResponse(response, ErrorCode.INVALID_ACCESS_TOKEN);
+        }
+        return; // 필터 체인 진행하지 않음
+      } else if (isAdminPageRequest) {
+        // 관리자 페이지: 로그인 페이지로 리다이렉트
+        log.error("관리자 페이지 요청 시, 토큰이 없거나 유효하지 않습니다.");
+        response.sendRedirect("/login");
+        return;
+      }
+    } catch (ExpiredJwtException e) {
+      log.error("토큰 만료: {}", e.getMessage());
+      // 토큰 만료 예외 처리
+      if (isApiRequest) {
+        sendErrorResponse(response, ErrorCode.EXPIRED_ACCESS_TOKEN);
+      } else {
+        response.sendRedirect("/login");
+      }
+      return;
     }
 
-    /**
-     * 에러 응답을 JSON 형태로 클라이언트에 전송
-     *
-     * @param response  HttpServletResponse 객체
-     * @param errorCode 발생한 에러코드
-     * @throws IOException
-     */
-    private void sendErrorResponse(HttpServletResponse response, ErrorCode errorCode) throws IOException {
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(errorCode.getStatus().value());
-        response.setCharacterEncoding("UTF-8");
+    // 필터 체인 계속 진행
+    filterChain.doFilter(request, response);
+  }
 
-        ErrorResponse errorResponse = new ErrorResponse(errorCode, errorCode.getMessage());
+  /**
+   * 에러 응답을 JSON 형태로 클라이언트에 전송
+   *
+   * @param response  HttpServletResponse 객체
+   * @param errorCode 발생한 에러코드
+   * @throws IOException
+   */
+  private void sendErrorResponse(HttpServletResponse response, ErrorCode errorCode) throws IOException {
+    response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+    response.setStatus(errorCode.getStatus().value());
+    response.setCharacterEncoding("UTF-8");
 
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.writeValue(response.getWriter(), errorResponse);
-    }
+    ErrorResponse errorResponse = new ErrorResponse(errorCode, errorCode.getMessage());
 
-    /**
-     * 화이트리스트 경로 확인 (인증x)
-     *
-     * @param uri 요청된 URI
-     * @return 화이트리스트 여부
-     */
-    private boolean isWhitelistedPath(String uri) {
-        return SecurityUrls.AUTH_WHITELIST.stream()
-                .anyMatch(pattern -> pathMatcher.match(pattern, uri));
-    }
+    ObjectMapper mapper = new ObjectMapper();
+    mapper.writeValue(response.getWriter(), errorResponse);
+  }
 
-    /**
-     * 관리자 권한 확인
-     *
-     * @param authentication 인증 정보
-     * @return 관리자 권한 여부
-     */
-    private boolean hasAdminRole(Authentication authentication) {
-        return authentication.getAuthorities().stream()
-                .anyMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"));
-    }
+  /**
+   * 화이트리스트 경로 확인 (인증x)
+   *
+   * @param uri 요청된 URI
+   * @return 화이트리스트 여부
+   */
+  private boolean isWhitelistedPath(String uri) {
+    return SecurityUrls.AUTH_WHITELIST.stream()
+        .anyMatch(pattern -> pathMatcher.match(pattern, uri));
+  }
+
+  /**
+   * 관리자 권한 확인
+   *
+   * @param authentication 인증 정보
+   * @return 관리자 권한 여부
+   */
+  private boolean hasAdminRole(Authentication authentication) {
+    return authentication.getAuthorities().stream()
+        .anyMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"));
+  }
 }
